+++ aprsd/dev.py
@@ -14,7 +14,9 @@ import click_completion
 
 # local imports here
 import aprsd
-from aprsd import client, plugin, utils
+from aprsd import client
+from aprsd import config as aprsd_config
+from aprsd import plugin
 
 
 # setup the global logger
@@ -156,7 +158,7 @@ def setup_logging(config, loglevel, quiet):
     "--config",
     "config_file",
     show_default=True,
-    default=utils.DEFAULT_CONFIG_FILE,
+    default=aprsd_config.DEFAULT_CONFIG_FILE,
     help="The aprsd config file to use for options.",
 )
 @click.option(
@@ -178,7 +180,7 @@ def test_plugin(
 ):
     """APRSD Plugin test app."""
 
-    config = utils.parse_config(config_file)
+    config = aprsd_config.parse_config(config_file)
 
     setup_logging(config, loglevel, False)
     LOG.info(f"Test APRSD PLugin version: {aprsd.__version__}")
