+++ aprsd/plugins/weather.py
@@ -4,7 +4,7 @@ import re
 
 import requests
 
-from aprsd import plugin, plugin_utils, trace, utils
+from aprsd import config, plugin, plugin_utils, trace
 
 
 LOG = logging.getLogger("APRSD")
@@ -34,7 +34,7 @@ class USWeatherPlugin(plugin.APRSDRegexCommandPluginBase):
         # message = packet.get("message_text", None)
         # ack = packet.get("msgNo", "0")
         try:
-            utils.check_config_option(self.config, ["services", "aprs.fi", "apiKey"])
+            config.check_config_option(self.config, ["services", "aprs.fi", "apiKey"])
         except Exception as ex:
             LOG.error(f"Failed to find config aprs.fi:apikey {ex}")
             return "No aprs.fi apikey found"
@@ -115,7 +115,7 @@ class USMetarPlugin(plugin.APRSDRegexCommandPluginBase):
             fromcall = fromcall
 
             try:
-                utils.check_config_option(
+                config.check_config_option(
                     self.config,
                     ["services", "aprs.fi", "apiKey"],
                 )
@@ -199,7 +199,7 @@ class OWMWeatherPlugin(plugin.APRSDRegexCommandPluginBase):
             searchcall = fromcall
 
         try:
-            utils.check_config_option(self.config, ["services", "aprs.fi", "apiKey"])
+            config.check_config_option(self.config, ["services", "aprs.fi", "apiKey"])
         except Exception as ex:
             LOG.error(f"Failed to find config aprs.fi:apikey {ex}")
             return "No aprs.fi apikey found"
@@ -220,7 +220,7 @@ class OWMWeatherPlugin(plugin.APRSDRegexCommandPluginBase):
         lon = aprs_data["entries"][0]["lng"]
 
         try:
-            utils.check_config_option(
+            config.check_config_option(
                 self.config,
                 ["services", "openweathermap", "apiKey"],
             )
@@ -229,7 +229,7 @@ class OWMWeatherPlugin(plugin.APRSDRegexCommandPluginBase):
             return "No openweathermap apiKey found"
 
         try:
-            utils.check_config_option(self.config, ["aprsd", "units"])
+            config.check_config_option(self.config, ["aprsd", "units"])
         except Exception:
             LOG.debug("Couldn't find untis in aprsd:services:units")
             units = "metric"
@@ -323,7 +323,7 @@ class AVWXWeatherPlugin(plugin.APRSDRegexCommandPluginBase):
             searchcall = fromcall
 
         try:
-            utils.check_config_option(self.config, ["services", "aprs.fi", "apiKey"])
+            config.check_config_option(self.config, ["services", "aprs.fi", "apiKey"])
         except Exception as ex:
             LOG.error(f"Failed to find config aprs.fi:apikey {ex}")
             return "No aprs.fi apikey found"
@@ -344,13 +344,13 @@ class AVWXWeatherPlugin(plugin.APRSDRegexCommandPluginBase):
         lon = aprs_data["entries"][0]["lng"]
 
         try:
-            utils.check_config_option(self.config, ["services", "avwx", "apiKey"])
+            config.check_config_option(self.config, ["services", "avwx", "apiKey"])
         except Exception as ex:
             LOG.error(f"Failed to find config avwx:apiKey {ex}")
             return "No avwx apiKey found"
 
         try:
-            utils.check_config_option(self.config, ["services", "avwx", "base_url"])
+            config.check_config_option(self.config, ["services", "avwx", "base_url"])
         except Exception as ex:
             LOG.debug(f"Didn't find avwx:base_url {ex}")
             base_url = "https://avwx.rest"
